Add re-centering of player character
I don't quite see the purpose here to be honest, but the apparent
intention is to make sure that the player "doesn't drift off".
Since he isn't actually moving, this does appear to be a weird to me.

Anyway, it doesn't place the player in the center of the platform
after rotating (since the platforms are moving themselves), so it
seems kind of pointless.

Anyway, this commit also adds spawning more platforms into the future
after leaving a T-section.