Add fog to world
Note that fog and Skyboxes don't go together well. I adjusted the
fog color to be close to the dark skybox background which works fine, as
long as we're not running towards the bright part of the skybox.
It may make sense to flip the skybox such that the bright part is
beneath or above us though.